TOY-57 copyFileToDownloads를 ImageUtils로 이동 및 적용
